requirements.txt
base58==0.2.4
boto3==1.4.1
botocore==1.4.80
click==6.6
Django==1.10.4
django-storages==1.5.1
djangorestframework==3.5.3
docutils==0.12
futures==3.0.5
hjson==2.0.2
jmespath==0.9.0
kappa==0.6.0
lambda-packages==0.10.0
placebo==0.8.1
python-dateutil==2.6.0
python-slugify==1.2.1
PyYAML==3.12
requests==2.12.3
s3transfer==0.1.9
six==1.10.0
tqdm==4.10.0
troposphere==1.9.0
Unidecode==0.4.19
Werkzeug==0.11.11
wheel==0.29.0
wsgi-request-logger==0.4.6
zappa==0.31.0
